Refactor `models.go` to extract components to their own files
`models.go` was getting large and unwieldy. Separating out the multiple
components that were in that file should make this easier to maintain in the
future.
