+++ models/user_data.go
@@ -1,0 +1,10 @@
+package models
+
+import "github.com/jinzhu/gorm"
+
+type UserData struct {
+	gorm.Model
+	Email string `gorm:"not null"`
+	Reg   []byte
+	Key   []byte
+}
